requirements-docker.txt
# Core dependencies
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0

# Audio processing
PyAudio==0.2.14
pyttsx3==2.98
openai-whisper==20240930

# Language models and chains
langchain==0.3.19
langchain-community==0.3.17
langchain-core==0.3.36
langchain-text-splitters==0.3.6
ollama==0.4.7

# Data handling and validation
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
coverage==7.6.12

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
PyYAML==6.0.2
